% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bim_payer.R
\name{bim_payer_cadth}
\alias{bim_payer_cadth}
\title{CADTH Canadian public payer perspective}
\usage{
bim_payer_cadth()
}
\value{
A \code{bim_payer} object.
}
\description{
Returns a \code{\link[=bim_payer]{bim_payer()}} representing the Canadian Drug Review (CDR)
public payer perspective used in CADTH submissions.
}
