% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bim_payer.R
\name{bim_payer_default}
\alias{bim_payer_default}
\title{Default payer perspective (healthcare system, 100\% coverage)}
\usage{
bim_payer_default()
}
\value{
A \code{bim_payer} object.
}
\description{
Returns a \code{\link[=bim_payer]{bim_payer()}} representing a generic healthcare system
perspective with full cost coverage. Used as the default in \code{\link[=bim_model]{bim_model()}}.
}
