% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bim_payer.R
\name{bim_payer_nhs}
\alias{bim_payer_nhs}
\title{NHS England payer perspective}
\usage{
bim_payer_nhs()
}
\value{
A \code{bim_payer} object.
}
\description{
Returns a \code{\link[=bim_payer]{bim_payer()}} representing the NHS England perspective,
appropriate for NICE Technology Appraisal submissions.
}
