% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bim_payer.R
\name{bim_payer_us_commercial}
\alias{bim_payer_us_commercial}
\title{US commercial payer perspective}
\usage{
bim_payer_us_commercial()
}
\value{
A \code{bim_payer} object.
}
\description{
Returns a \code{\link[=bim_payer]{bim_payer()}} representing a US commercial insurer perspective.
}
