% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bim_payer.R
\name{print.bim_payer}
\alias{print.bim_payer}
\title{Print method for bim_payer}
\usage{
\method{print}{bim_payer}(x, ...)
}
\arguments{
\item{x}{A \code{bim_payer} object.}

\item{...}{Further arguments (ignored).}
}
\description{
Print method for bim_payer
}
