% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bim_model.R
\name{summary.bim_model}
\alias{summary.bim_model}
\title{Summary method for bim_model}
\usage{
\method{summary}{bim_model}(object, digits = 0L, ...)
}
\arguments{
\item{object}{A \code{bim_model} object.}

\item{digits}{\code{integer(1)}. Decimal digits for currency amounts. Default \code{0}.}

\item{...}{Further arguments (ignored).}
}
\value{
The \code{bim_model} object, invisibly.
}
\description{
Summary method for bim_model
}
