% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bim_population.R
\name{summary.bim_population}
\alias{summary.bim_population}
\title{Summary method for bim_population}
\usage{
\method{summary}{bim_population}(object, ...)
}
\arguments{
\item{object}{A \code{bim_population} object.}

\item{...}{Further arguments (ignored).}
}
\value{
The \code{bim_population} object, invisibly.
}
\description{
Summary method for bim_population
}
