% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bim_costs.R
\name{print.bim_costs}
\alias{print.bim_costs}
\title{Print method for bim_costs}
\usage{
\method{print}{bim_costs}(x, ...)
}
\arguments{
\item{x}{A \code{bim_costs} object.}

\item{...}{Further arguments (ignored).}
}
\description{
Print method for bim_costs
}
