% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bim_sensitivity.R
\name{print.bim_dsa}
\alias{print.bim_dsa}
\title{Print method for bim_dsa}
\usage{
\method{print}{bim_dsa}(x, ...)
}
\arguments{
\item{x}{A \code{bim_dsa} object.}

\item{...}{Further arguments (ignored).}
}
\description{
Print method for bim_dsa
}
